Add WASM SIMD optimizations and cross-platform f32 tolerant snapshot tests
- Enable simd128 target feature for wasm32 and add --enable-simd to wasm build pipeline
- Add SIMD-accelerated PCM format conversions (mono↔stereo, 8↔16bit) in pcm.rs
- Optimize tone_sample() with combined velocity/volume/pan multiply chain and 2-channel parallel processing
- Add safe-batch moo loop that skips per-sample event checks between event boundaries
- Change oscillator synthesis to f32 for faster sin() and better SIMD potential
- Update snapshot comparison to allow ±2 i16 tolerance to accommodate f32::sin() differences between Apple libm and V8/Deno libm

+++ src/pulse/noise_builder.rs
@@ -156,7 +156,7 @@ impl NoiseBuilder {
       WaveType::Sine => {
         osci.ready_get_sample(vec![Point { x: 1, y: 128 }], 128, SMP_COUNT as u32, 0);
         (0..SMP_COUNT as u32)
-          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f64) as i16)
+          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f32) as i16)
           .collect()
       }
       WaveType::Saw => {
@@ -189,7 +189,7 @@ impl NoiseBuilder {
           0,
         );
         (0..SMP_COUNT as u32)
-          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f64) as i16)
+          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f32) as i16)
           .collect()
       }
       WaveType::Rect2 => {
@@ -205,7 +205,7 @@ impl NoiseBuilder {
           0,
         );
         (0..SMP_COUNT as u32)
-          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f64) as i16)
+          .map(|s| (osci.get_one_sample_overtone(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f32) as i16)
           .collect()
       }
       WaveType::Tri => {
@@ -226,7 +226,7 @@ impl NoiseBuilder {
         );
         (0..SMP_COUNT as u32)
           .map(|s| {
-            (osci.get_one_sample_coordinate(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f64) as i16
+            (osci.get_one_sample_coordinate(s).clamp(-1.0, 1.0) * SAMPLING_TOP as f32) as i16
           })
           .collect()
       }
